requirements.txt
pyspark==3.5.2
requests==2.32.3
deltalake==0.19.2
delta-spark==3.2.0
ruff==0.6.5
PyYAML==6.0.2
urllib3==2.2.3
duckdb==1.2.0
pandas==2.2.3